start pulling non-pygments things out of pyghooks
